Finished version of projective transform. The main bug was dealing with the fact that matrices are indexed by (row, col) but all the math for figuring out bounding box and inverse-mapped points is done with (x, y), where (x -> col, y -> row).
There was also a problem with leaving a byte in the stream before
reading the rest of a binary image in both Pgm and Ppm. In Ppm this was
more apparent because all Vector3b values would be shifted by one, but
in a Pgm image this was not as obvious.